+++ src/layout/routerView/iframes.vue
@@ -1,11 +1,11 @@
 <template>
-	<div class="layout-view-bg-white flex" :style="{ height: `calc(100vh - ${iframeHeight}`, border: 'none' }" v-loading="iframeLoading">
+	<div class="layout-view-bg-white flex" :style="{ height: `calc(100vh - ${setIframeHeight}`, border: 'none' }" v-loading="iframeLoading">
 		<iframe :src="iframeUrl" frameborder="0" height="100%" width="100%" id="iframe" v-show="!iframeLoading"></iframe>
 	</div>
 </template>
 
 <script lang="ts">
-import { defineComponent, reactive, toRefs, onMounted, nextTick, watch } from 'vue';
+import { defineComponent, reactive, toRefs, onMounted, nextTick, watch, computed } from 'vue';
 import { useRoute } from 'vue-router';
 import { useStore } from '/@/store/index';
 export default defineComponent({
@@ -16,7 +16,6 @@ export default defineComponent({
 		const state = reactive({
 			iframeLoading: true,
 			iframeUrl: '',
-			iframeHeight: '',
 		});
 		// 初始化页面加载 loading
 		const initIframeLoad = () => {
@@ -31,15 +30,14 @@ export default defineComponent({
 			});
 		};
 		// 设置 iframe 的高度
-		const initIframeHeight = () => {
+		const setIframeHeight = computed(() => {
 			let { isTagsview } = store.state.themeConfig.themeConfig;
-			if (isTagsview) return (state.iframeHeight = `84px`);
-			else return (state.iframeHeight = `20px`);
-		};
+			if (isTagsview) return `84px`;
+			else return `50px`;
+		});
 		// 页面加载时
 		onMounted(() => {
 			initIframeLoad();
-			initIframeHeight();
 		});
 		// 监听路由变化，多个 iframe 时使用
 		watch(
@@ -49,6 +47,7 @@ export default defineComponent({
 			}
 		);
 		return {
+			setIframeHeight,
 			...toRefs(state),
 		};
 	},