+++ src/layout/routerView/link.vue
@@ -1,5 +1,5 @@
 <template>
-	<div class="layout-view-bg-white flex layout-view-link" :style="{ height: `calc(100vh - ${linkHeight}` }">
+	<div class="layout-view-bg-white flex layout-view-link" :style="{ height: `calc(100vh - ${setLinkHeight}` }">
 		<a href="https://element-plus.gitee.io/#/zh-CN/component/installation" target="_blank" class="flex-margin"
 			>{{ currentRouteMeta.title }}：{{ currentRouteMeta.isLink }}</a
 		>
@@ -7,7 +7,7 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, toRefs, reactive, onMounted } from 'vue';
+import { defineComponent, toRefs, reactive, onMounted, computed } from 'vue';
 import { useRoute } from 'vue-router';
 import { useStore } from '/@/store/index';
 export default defineComponent({
@@ -17,24 +17,23 @@ export default defineComponent({
 		const store = useStore();
 		const state = reactive({
 			currentRouteMeta: {},
-			linkHeight: '',
 		});
 		// 初始化获取当前路由 meta
 		const initGetMeta = () => {
 			state.currentRouteMeta = route.meta;
 		};
-		// 设置 iframe 的高度
-		const initLinkHeight = () => {
+		// 设置 link 的高度
+		const setLinkHeight = computed(() => {
 			let { isTagsview } = store.state.themeConfig.themeConfig;
-			if (isTagsview) return (state.linkHeight = `114px`);
-			else return (state.linkHeight = `50px`);
-		};
+			if (isTagsview) return `114px`;
+			else return `80px`;
+		});
 		// 页面加载时
 		onMounted(() => {
 			initGetMeta();
-			initLinkHeight();
 		});
 		return {
+			setLinkHeight,
 			...toRefs(state),
 		};
 	},