+++ src/utils/setIconfont.ts
@@ -1,6 +1,6 @@
 // 字体图标 url
 const cssCdnUrlList: Array<string> = [
-	'//at.alicdn.com/t/font_2298093_ke79kmw5s7r.css',
+	'//at.alicdn.com/t/font_2298093_h8g6cpatww5.css',
 	'//netdna.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css',
 ];
 // 第三方 js url