+++ src/views/pages/listAdapt/mock.ts
@@ -8,7 +8,6 @@ export const filterList = [
 		price: (Math.random() * 10).toFixed(2),
 		monSales: (Math.random() * 20).toFixed(2),
 		id: 1,
-		isShowMore: false,
 	},
 	{
 		img: 'http://news.sznews.com/pic/2021-03/09/78cf72b6-e2d9-459d-a368-470414a027f4679cf4ea-26fa-48c8-9fee-c2d092a91400.png',
@@ -18,7 +17,6 @@ export const filterList = [
 		price: (Math.random() * 10).toFixed(2),
 		monSales: (Math.random() * 20).toFixed(2),
 		id: 2,
-		isShowMore: false,
 	},
 	{
 		img: 'http://news.sznews.com/pic/2021-03/09/1faf3c6e-1250-4e6b-b072-4a331553e027.jpg',
@@ -28,7 +26,6 @@ export const filterList = [
 		price: (Math.random() * 10).toFixed(2),
 		monSales: (Math.random() * 20).toFixed(2),
 		id: 3,
-		isShowMore: false,
 	},
 	{
 		img: 'https://news.sznews.com/pic/2021-03/09/9fcf6dd4-1e80-4497-bdc9-83dc7246d170.jpg.2',
@@ -38,7 +35,6 @@ export const filterList = [
 		price: (Math.random() * 10).toFixed(2),
 		monSales: (Math.random() * 20).toFixed(2),
 		id: 4,
-		isShowMore: false,
 	},
 	{
 		img: 'https://news.sznews.com/pic/2021-03/09/1bd78227-4126-4a43-bdf6-48ead6edd1bf.jpg.2',
@@ -48,7 +44,6 @@ export const filterList = [
 		price: (Math.random() * 10).toFixed(2),
 		monSales: (Math.random() * 20).toFixed(2),
 		id: 5,
-		isShowMore: false,
 	},
 	{
 		img: 'http://news.sznews.com/pic/2021-03/08/9ea943a3-3ae8-4f49-8296-711ec36ef8c6_watermark.png',
@@ -58,7 +53,6 @@ export const filterList = [
 		price: (Math.random() * 10).toFixed(2),
 		monSales: (Math.random() * 20).toFixed(2),
 		id: 6,
-		isShowMore: false,
 	},
 	{
 		img: 'https://news.sznews.com/pic/2021-03/08/a95ba232-1422-4f7e-b85f-c61d486c8659.jpg.2',
@@ -68,7 +62,6 @@ export const filterList = [
 		price: (Math.random() * 10).toFixed(2),
 		monSales: (Math.random() * 20).toFixed(2),
 		id: 7,
-		isShowMore: false,
 	},
 	{
 		img: 'http://news.sznews.com/pic/2021-03/08/76816bf0-3899-4c7e-bc6e-079b5ba8725e.jpg',
@@ -78,7 +71,6 @@ export const filterList = [
 		price: (Math.random() * 10).toFixed(2),
 		monSales: (Math.random() * 20).toFixed(2),
 		id: 8,
-		isShowMore: false,
 	},
 	{
 		img: 'https://news.sznews.com/pic/2021-03/08/28ed70d4-71f5-4abb-bf7b-0294bece9e43.jpg.2',
@@ -88,7 +80,6 @@ export const filterList = [
 		price: (Math.random() * 10).toFixed(2),
 		monSales: (Math.random() * 20).toFixed(2),
 		id: 9,
-		isShowMore: false,
 	},
 	{
 		img: 'http://news.sznews.com/pic/2021-03/05/d13ae31f-fd45-431a-b48e-c5895bbc193e.png',
@@ -98,6 +89,5 @@ export const filterList = [
 		price: (Math.random() * 10).toFixed(2),
 		monSales: (Math.random() * 20).toFixed(2),
 		id: 10,
-		isShowMore: false,
 	},
 ];